requirements.txt
flask==3.0.2
tabula-py==2.9.0
pandas==2.2.1
openpyxl==3.1.2
pillow==10.2.0
pystray==0.19.5
pyinstaller==6.13.0 